fix: lineaEstimateGas refactoring
