--- src/linea/types/fee.ts
@@ -1,11 +1,0 @@
-import type { EstimateFeesPerGasReturnType } from '../../index.js'
-
-export type LineaEstimateGasReturnType = {
-  gasLimit: bigint
-  baseFeePerGas: bigint
-  priorityFeePerGas: bigint
-}
-
-export type LineaEstimateFeesPerGasReturnType = {
-  gasLimit: bigint
-} & EstimateFeesPerGasReturnType
